extract toolbox-button
